completed FileSystem.sync completed FileSystem.delete completed FileSystem.fsize completed FileSystem.seek completed FileSystem.deallocateBlocks
TODO: read, write, open, close, format

+++ src/Directory.java
@@ -50,7 +50,7 @@ public class Directory {
 	}
 
 	// returns true, given 1
-	public boolean ifree(short iNumber) {
+	public boolean ifree(int iNumber) {
 		return false;
 		// deallocates this inumber (inode number)
 		// the corresponding file will be deleted.
